--- packages/website/src/pageNavigation.ts
@@ -1,6 +1,0 @@
-export const pageNavigation = [
-  { href: '/', children: 'Strona główna' },
-  { href: '/mentor', children: 'Dla mentorów' },
-  { href: '/faq', children: 'FAQ' },
-  { href: '/kontakt', children: 'Kontakt' },
-];
